completely get rid of django_cherrypy_wsgiserver by moving its only active component to kalite.distributed
